backends/backend-flask/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
marshmallow==3.20.1
pika==1.3.2
pytest==7.4.3
psycopg2-binary==2.9.9
prometheus-flask-exporter==0.23.0
redis==5.0.1
gunicorn==21.2.0
flask-caching==2.1.0
